requirements.txt
absl-py==2.3.1
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
autoawq==0.2.9
autograd==1.8.0
autograd-gamma==0.5.0
beautifulsoup4==4.13.4
bertopic==0.17.0
bitsandbytes==0.47.0
blake3==1.0.5
blessed==1.17.12
bs4==0.0.2
cachetools==6.1.0
cbor2==5.7.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
compressed-tensors==0.10.2
contourpy==1.3.2
cryptography==45.0.5
cuda-bindings==12.9.1
cuda-pathfinder==1.1.0
cuda-python==12.9.0
cudf-cu12==25.8.0
cuml-cu12==25.8.0
cupy-cuda12x==13.5.1
cuvs-cu12==25.8.0
cycler==0.12.1
dask==2025.7.0
dask-cuda==25.8.0
dask-cudf-cu12==25.8.0
datasets==3.6.0
decorator==5.2.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distributed==2025.7.0
distributed-ucxx-cu12==0.45.1
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email-validator==2.2.0
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.18.0
flash-attn==2.8.1
flashinfer-python==0.2.11.post3
fonttools==4.58.5
formulaic==1.1.1
frozenlist==1.7.0
fsspec==2025.7.0
gguf==0.17.1
gitdb==4.0.12
gitpython==3.1.44
graphviz==0.21
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.7
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib-metadata==8.7.0
interegular==0.3.3
interface-meta==1.3.0
ipdb==0.13.13
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
levenshtein==0.27.1
libcudf-cu12==25.8.0
libcuml-cu12==25.8.0
libcuvs-cu12==25.8.0
libkvikio-cu12==25.8.0
libraft-cu12==25.8.0
librmm-cu12==25.8.0
libucx-cu12==1.18.1
libucxx-cu12==0.45.1
lifelines==0.30.0
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
locket==1.0.0
mako==1.3.10
markdown==3.8.2
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral-common==1.8.3
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
narwhals==1.41.0
networkx==3.5
ninja==1.13.0
numba==0.61.2
numba-cuda==0.14.0
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cudnn-frontend==1.14.0
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvtx==0.2.13
objexplore==1.6.3
objprint==0.3.0
openai==1.90.0
opencv-python-headless==4.12.0.88
optuna==4.3.0
outlines-core==0.2.10
packaging==25.0
pandas==2.3.1
parso==0.8.4
partd==1.4.2
partial-json-parser==0.2.1.1.post6
peft==0.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.1.2
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pygments==2.19.2
pylibcudf-cu12==25.8.0
pylibraft-cu12==25.8.0
pynndescent==0.5.13
pynvjitlink-cu12==0.7.0
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.1
raft-dask-cu12==25.8.0
rapidfuzz==3.13.0
rapids-dask-dependency==25.8.0
rapids-logger==0.1.1
ray==2.48.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rmm-cu12==25.8.0
rpds-py==0.27.0
safetensors==0.6.2
scikit-learn==1.7.0
scipy==1.16.1
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.1
sentry-sdk==2.34.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
sqlalchemy==2.0.41
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
tblib==3.1.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thefuzz==0.22.1
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
toolz==1.0.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
treelite==4.4.1
triton==3.3.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
ucx-py-cu12==0.45.0
ucxx-cu12==0.45.1
umap-learn==0.5.9.post2
unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.10.0
wandb==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.31
xgrammar==0.1.21
xxhash==3.5.0
yarl==1.20.1
zict==3.0.0
zipp==3.23.0
zstandard==0.23.0
